refacttor: cansubmitEdit 함수 선언
